+++ plugins/blog/plugin-comment/src/client/config.ts
@@ -1,4 +1,3 @@
-import type { ClientConfig } from 'vuepress/client'
 import { defineClientConfig } from 'vuepress/client'
 import CommentService from './components/CommentService.js'
 import { injectCommentConfig } from './helpers/index.js'
@@ -11,4 +10,4 @@ export default defineClientConfig({
     injectCommentConfig(app)
     app.component('CommentService', CommentService)
   },
-}) as ClientConfig
+})