+++ themes/theme-default/src/client/composables/useRelatedLinks.ts
@@ -114,10 +114,10 @@ export const useRelatedLinks = (): RelatedLinks => {
 
     return prevConfig === false
       ? null
-      : prevConfig ??
+      : (prevConfig ??
           (themeLocale.value.prev === false
             ? null
-            : resolveFromSidebarItems(sidebarItems.value, route.path, -1))
+            : resolveFromSidebarItems(sidebarItems.value, route.path, -1)))
   })
 
   const nextLink = computed(() => {
@@ -128,10 +128,10 @@ export const useRelatedLinks = (): RelatedLinks => {
 
     return nextConfig === false
       ? null
-      : nextConfig ??
+      : (nextConfig ??
           (themeLocale.value.next === false
             ? null
-            : resolveFromSidebarItems(sidebarItems.value, route.path, 1))
+            : resolveFromSidebarItems(sidebarItems.value, route.path, 1)))
   })
 
   return {